Implementing different fusion techniques.
Implemented Attention mechanism for the final concatenated vector(Containing image and text features)
in order to capture the essence of correlations of features among different modalities. (YET TO RUN AND SEE THE RESULTS).
Also implemented GMU(Gated Multimodal Unit), where the text and the image feature are fed into 
our GMU block in turn containing a fusion and a gating layer to produce a more meaningful vector representation. ((YET TO RUN AND SEE THE RESULTS))